Reorganized files and started creating basic components
