+++ friends/src/App.js
@@ -1,11 +1,13 @@
 import React, { Component } from 'react';
-import './App.css';
+import './styles/App.css';
+
+import FriendsList from './components/FriendsList';
 
 class App extends Component {
   render() {
     return (
       <div className="App">
-        <h1>testing</h1>
+        <FriendsList />
       </div>
     );
   }
