+++ friends/src/components/Friend.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Friend = (props) => {
+  return (
+    <div>
+      <h1>Friend</h1>
+    </div>
+  );
+};
+
+export default Friend;
