+++ friends/src/components/FriendsList.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import Friend from './Friend';
+
+const FriendsList = (props) => {
+  return (
+    <div>
+      <h1>Friend List</h1>
+      <Friend />
+    </div>
+  );
+};
+
+export default FriendsList;
